Moved setup and teardwon related classes and objects from DependencyExtension_ParentChain. Made ScopeSubtree tests an internal inheritor of it.
