+++ Assets/UnityDependency/Tests/Editor/CaptureScopeTests/DependencyExtension_ParentChain.cs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 40ad61089f8e79e4cb932191e4dacd33
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
