+++ Assets/UnityDependency/Tests/Editor/CaptureScopeTests/ScopeSubtree.cs
@@ -1,0 +1,59 @@
+﻿namespace UnityDependency.Test.CaptureScope
+{
+    using Alquimiaware;
+    using NUnit.Framework;
+    using UnityEngine;
+    using UnityHelpers;
+
+    public abstract partial class DependencyExtension_ParentChain
+    {
+        [TestFixture]
+        public class ScopeSubtree : DependencyExtension_ParentChain
+        {
+            [Test]
+            public void Capture_CandidateInScene_IsNotCaptured()
+            {
+                var candidate = this.goBuilder.CreateInRoot<BoxCollider>("Candidate");
+
+                this.sut.CaptureDependencies();
+                Assert.AreNotSame(sut.subtreeCollider, candidate);
+            }
+
+            [Test]
+            public void Capture_CandidateInChild_IsCaptured()
+            {
+                var candidate = this.parentChain.Child.AddComponent<BoxCollider>();
+
+                this.sut.CaptureDependencies();
+                Assert.AreSame(sut.subtreeCollider, candidate);
+            }
+
+            [Test]
+            public void Capture_CandidateInGrandchild_IsCaptured()
+            {
+                var candidate = this.parentChain.Grandchild.AddComponent<BoxCollider>();
+
+                this.sut.CaptureDependencies();
+                Assert.AreSame(sut.subtreeCollider, candidate);
+            }
+
+            [Test]
+            public void Capture_CandidateInSelf_IsCaptured()
+            {
+                var candidate = this.parentChain.Self.AddComponent<BoxCollider>();
+
+                this.sut.CaptureDependencies();
+                Assert.AreSame(sut.subtreeCollider, candidate);
+            }
+
+            [Test]
+            public void Capture_CandidateInParent_IsNotCaptured()
+            {
+                var candidate = this.parentChain.Parent.AddComponent<BoxCollider>();
+
+                this.sut.CaptureDependencies();
+                Assert.AreNotSame(sut.subtreeCollider, candidate);
+            }
+        }
+    }
+}
